make server routes moudular
