+++ public/scripts/app.js
@@ -1,14 +1,4 @@
 $(() => {
-  $.ajax({
-    method: 'GET',
-    url: '/api/users'
-  }).done(users => {
-    for (user of users) {
-      $('<div>')
-        .text(user.name)
-        .appendTo($('body'));
-    }
-  });
 
   function resetCounter() {
     $('label.counterText').each(function(indivCounterText, val) {
@@ -71,12 +61,12 @@ $(() => {
     });
     // $.post("/vote", {option: optionArray, randomURL: randomURL});
     $.ajax({
-      url: '/vote',
+      url: '/api/users/vote',
       type: 'POST',
       data: {option: optionArray, randomURL: randomURL},
       success:function(result){
         console.log("we are good in the succcess function",result);
-        window.location = "/thanks";
+        window.location = "/api/users/thanks";
       },
       error: function(error){
         console.log("we are in the error code");
