+++ views/admin.ejs
@@ -23,7 +23,7 @@
         <% };%>
       </ol>
 
-      <h5>This is the voting <a href="/user/<%=randomURL%>">link</a></h5>
+      <h5>This is the voting <a href="/api/users/user/<%=randomURL%>">link</a></h5>
   </main>
   </div>
 </body>
