+++ views/index.ejs
@@ -10,7 +10,7 @@
   <% include partials/_header %>
 
   <main class="container">
-    <form action="/create" method="post">
+    <form action="/api/users/create" method="post">
       <h3>Create New Poll</h3>
       <div class="test">
         <div class="email">
